rename Pesto to Evaluator
